requirements.txt
Pillow
piexif
requests
